Separete CommitCompare class to commit_compare.rb file
close #6
